+++ kml.js
@@ -155,7 +155,11 @@ function csv(img, level) {
     t.push('NEUTRAL');
   }
 
-  var csvstr = '"Team","Level","Name","Address","Energy","Links","Mods","Resonators","PHOTO"\n';
+  var csvstr = '"Team","Level","Name","Address","Energy %","Links","Mods","Resonators",';
+  if (img) {
+    csvstr += "Photo URL";
+  }
+  csvstr += '\n';
 
   n.forEach(function(v){
     var idx = levels[v];
@@ -181,7 +185,10 @@ function csv(img, level) {
         csvstr += ',"'+(0+l.links)+'"';
         csvstr += ',"'+(0+l.mods)+'"';
         csvstr += ',"'+vxml(l.resonators.join('') || '-')+'"';
-        csvstr += ',"'+(l.imageUrl||'')+'"\n';
+        if (img) {
+          csvstr += ',"'+(l.imageUrl||'')+'"';
+        }
+        csvstr += '\n';
       }
     });
   });